Refactored SectorModel and tests to reflect changes
